变更NPC交互提示+铁匠交互细节+shift奔跑精进
1.使用blender制作了更好看的“任务可接取”的NPC头顶UI。
2.增加了铁匠专属的交互镜头
3.对于Enegy，使用float代替了int ，在精力不足15点时无法奔跑，现在不会出现可以无限奔跑的问题的（之后要优化，使得只要进入walk的速度就恢复精力，而不是现在跑完还要松开shift键才恢复）

+++ Combat System Course 2/Assets/Scripts/Player/PlayerProperty.cs
@@ -13,7 +13,7 @@ public class PlayerProperty : MonoBehaviour
     [SerializeField] private float sprintCostPerSecond = 15f;        // 冲刺每秒消耗
     [SerializeField] private int rollEnergyCost = 20;                // 每次翻滚消耗
 
-    public int EnergyValue => energyValue;                           // 只读对外暴露
+    public float EnergyValue => energyValue;                           // 只读对外暴露
     public int MaxEnergy => maxEnergy;
     public float EnergyNormalized => (float)energyValue / maxEnergy;
 
@@ -25,7 +25,7 @@ public class PlayerProperty : MonoBehaviour
     public static PlayerProperty Instance;
     public Dictionary<PropertyType, List<Property>> propertyDict;
     public int hpValue = 100;
-    public int energyValue = 100;
+    public float energyValue = 100;
     public int armorValue => baseArmorValue + equipmentArmorBonus; // 只读属性：总护甲值
     public int level = 1;
     public int currEXP = 0;
@@ -232,7 +232,7 @@ public class PlayerProperty : MonoBehaviour
         OnArmorChanged?.Invoke();
     }
     // 消耗能量（可被外部调用）
-    public bool ConsumeEnergy(int amount)
+    public bool ConsumeEnergy(float amount)
     {
         if (energyValue >= amount)
         {
@@ -241,11 +241,11 @@ public class PlayerProperty : MonoBehaviour
             OnEnergyChanged?.Invoke(EnergyNormalized);
             return true;
         }
-        return false; // 能量不足
+        return false;
     }
 
     // 恢复能量（内部用）
-    public void RestoreEnergy(int amount)
+    public void RestoreEnergy(float amount)
     {
         energyValue += amount;
         energyValue = Mathf.Clamp(energyValue, 0, maxEnergy);
